+++ CotEditor/Sources/Application/Defaults/DefaultKeys.swift
@@ -88,8 +88,6 @@ extension DefaultKeys {
     static let indentWithTabKey = DefaultKey<Bool>("indentWithTabKey")
     static let autoLinkDetection = DefaultKey<Bool>("autoLinkDetectionKey")
     static let checkSpellingAsType = DefaultKey<Bool>("checkSpellingAsType")
-    static let highlightBraces = DefaultKey<Bool>("highlightBraces")
-    static let highlightLtGt = DefaultKey<Bool>("highlightLtGt")
     static let highlightSelectionInstance = DefaultKey<Bool>("highlightSelectionInstance")
     static let selectionInstanceHighlightDelay = DefaultKey<Double>("selectionInstanceHighlightDelay")