+++ CotEditor/Sources/Settings Window/Panes/EditSettingsView.swift
@@ -42,8 +42,6 @@ struct EditSettingsView: View {
     @AppStorage(.trimsWhitespaceOnlyLines) private var trimsWhitespaceOnlyLines
     
     @AppStorage(.autoLinkDetection) private var autoLinkDetection
-    @AppStorage(.highlightBraces) private var highlightBraces
-    @AppStorage(.highlightLtGt) private var highlightLtGt
     @AppStorage(.highlightSelectionInstance) private var highlightSelectionInstance
     @AppStorage(.selectionInstanceHighlightDelay) private var selectionInstanceHighlightDelay
     
@@ -92,10 +90,6 @@ struct EditSettingsView: View {
                 
                 VStack(alignment: .leading, spacing: isLiquidGlass ? nil : 6) {
                     Toggle(String(localized: "Link URLs in document", table: "EditSettings"), isOn: $autoLinkDetection)
-                    Toggle(String(localized: "Highlight matching braces “()” “[]” “{}”", table: "EditSettings"), isOn: $highlightBraces)
-                    Toggle(String(localized: "Highlight “<>”", table: "EditSettings"), isOn: $highlightLtGt)
-                        .disabled(!self.highlightBraces)
-                        .padding(.leading, 20)
                     Toggle(String(localized: "Highlight instances of selected text", table: "EditSettings"), isOn: $highlightSelectionInstance)
                     HStack(alignment: .firstTextBaseline) {
                         Text("Delay:", tableName: "EditSettings")